int64

+++ domain/topic/usecase/topic_usecase_test.go
@@ -48,6 +48,6 @@ func TestFetchTopics(t *testing.T) {
 	result, pagination, err := uc.FetchTopics(context.TODO(), 10, 0)
 
 	assert.NoError(t, err)
-	assert.Equal(t, 5, pagination.NextCursor)
+	assert.Equal(t, int64(5), pagination.NextCursor)
 	assert.Len(t, result, len(result))
 }
